added dynamic WPS domain placement, refactored forecast.py

+++ src/vis/postprocessor.py
@@ -232,7 +232,8 @@ class Postprocessor(object):
         # cleanup
         if cleanup:
             os.remove(raster_path)
-            os.remove(cb_path)
+            if cb_path is not None:
+                os.remove(cb_path)
         
         return kmz_path, raster_path, cb_path, corner_coords
